# Speed up erode-border calculation