--- a/PyRoute/Star.py
+++ b/PyRoute/Star.py
@@ -143,8 +143,9 @@ class Star(object):
         belts = str(Utilities.int_to_ehex(self.belts))
         ggCount = str(Utilities.int_to_ehex(self.ggCount))
         worlds = self.worlds
+        basecode = str(self.baseCode).upper()
 
-        result += str(self.baseCode).ljust(2) + " " + str(self.zone).ljust(1) + " " + popM + belts + ggCount + " "
+        result += basecode.ljust(2) + " " + str(self.zone).ljust(1) + " " + popM + belts + ggCount + " "
         result += str(worlds).ljust(2) + " " + str(self.alg_code).ljust(4) + " "
         result += str(star_list).ljust(14) + " " + " ".join(self.routes).ljust(41)