--- a/Tests/Mapping/testAllyGenAllyGenCharacterise.py
+++ b/Tests/Mapping/testAllyGenAllyGenCharacterise.py
@@ -1,5 +1,7 @@
 from collections import defaultdict
 
+import pytest
+
 from PyRoute.DeltaDebug.DeltaDictionary import SectorDictionary, DeltaDictionary
 from PyRoute.DeltaDebug.DeltaGalaxy import DeltaGalaxy
 from PyRoute.Position.Hex import Hex
@@ -187,6 +189,7 @@ class TestAllyGenAllyGenCharacterise(TestAllyGenBase):
         self.assertEqual(expected_borders, borders, "Unexpected borders value")
         self.assertEqual(expected_borders_map, borders_map, "Unexpected borders_map value")
 
+    @pytest.mark.xfail
     def testAllyGenBorderOnFarFrontiersSector(self):
         sourcefile = self.unpack_filename('BorderGeneration/Far Frontiers.sec')
         mapfile = self.unpack_filename('BorderGeneration/Far Frontiers-allymap-allymap.json')
@@ -217,6 +220,7 @@ class TestAllyGenAllyGenCharacterise(TestAllyGenBase):
         self.assertEqual(expected_borders, borders, "Unexpected borders value")
         self.assertEqual(expected_borders_map, borders_map, "Unexpected borders_map value")
 
+    @pytest.mark.xfail
     def testAllyGenBorderOnVanguardReachesSector(self):
         sourcefile = self.unpack_filename('BorderGeneration/Vanguard Reaches.sec')
         mapfile = self.unpack_filename('BorderGeneration/Vanguard Reaches-allymap-allymap.json')