--- a/pytest.ini
+++ b/pytest.ini
@@ -8,6 +8,7 @@ testpaths =
     Tests/Hypothesis/Inputs/**.py
     Tests/Hypothesis/Position/**.py
     Tests/Inputs/**.py
+    Tests/Mapping/**.py
     Tests/Outputs/**.py
     Tests/Pathfinding/**.py
     Tests/Position/**.py