… Tests/BorderGeneration/Far Frontiers-allymap-allymap.json
… Tests/BorderGeneration/Far Frontiers-allymap-border.json
… Tests/BorderGeneration/Far Frontiers-allymap-bordermap.json
… Tests/BorderGeneration/Far Frontiers-erode-allymap.json
… Tests/BorderGeneration/Far Frontiers-erode-border.json
… Tests/BorderGeneration/Far Frontiers-erode-bordermap.json
… Tests/BorderGeneration/Far Frontiers-range-allymap.json
… Tests/BorderGeneration/Far Frontiers-range-border.json
… Tests/BorderGeneration/Far Frontiers-range-bordermap.json
… Tests/BorderGeneration/Far Frontiers.sec
… Tests/BorderGeneration/Vanguard Reaches-allymap-allymap.json
… Tests/BorderGeneration/Vanguard Reaches-allymap-border.json
… Tests/BorderGeneration/Vanguard Reaches-allymap-bordermap.json
… Tests/BorderGeneration/Vanguard Reaches-erode-allymap.json
… Tests/BorderGeneration/Vanguard Reaches-erode-border.json
… Tests/BorderGeneration/Vanguard Reaches-erode-bordermap.json
… Tests/BorderGeneration/Vanguard Reaches-range-allymap.json
… Tests/BorderGeneration/Vanguard Reaches-range-border.json
… Tests/BorderGeneration/Vanguard Reaches-range-bordermap.json
… Tests/BorderGeneration/Vanguard Reaches.sec
… Tests/Mapping/testAllyGenBase.py
… Tests/Mapping/testAllyGenErodeCharacterise.py
… Tests/Mapping/testAllyGenRangeCharacterise.py