I can't impeach the speedup I managed over all of charted space.  The regression tests I added keep passing.  Maybe the extra 30x speedup is legit.

status quo, full-orchestra:
2024-04-03 08:53:15,008 - INFO - setting borders...
2024-04-03 08:53:15,009 - INFO - Processing worlds for erode map drawing
2024-04-03 11:56:42,485 - INFO - sorting routes...

3 h 2m 27s - 10947s

with improved erode-border speedups, full-orchestra, no route gen:
2024-04-05 22:38:46,386 - INFO - setting borders...
2024-04-05 22:38:46,386 - INFO - Processing worlds for erode map drawing
2024-04-05 22:40:35,411 - INFO - Processing ETI for worlds

1m 51s - 98.6x

status-quo, implist:
2024-04-05 00:02:21,711 - INFO - setting borders...
2024-04-05 00:02:21,711 - INFO - Processing worlds for erode map drawing
2024-04-05 00:03:42,339 - INFO - Processing ETI for worlds

1m 21s - 81s

with erode-border-speedups, implist, no route gen:
2024-04-04 23:59:13,048 - INFO - setting borders...
2024-04-04 23:59:13,048 - INFO - Processing worlds for erode map drawing
2024-04-04 23:59:45,396 - INFO - Processing ETI for worlds

32 s - 2.5x

status quo, over the zhodani heartland:
2024-04-04 22:31:47,208 - INFO - setting borders...
2024-04-04 22:31:47,208 - INFO - Processing worlds for erode map drawing
2024-04-04 22:31:58,862 - INFO - sorting routes...

11.6s

with erode-border speedups, over the zhodani heartland:
2024-04-04 22:29:17,756 - INFO - setting borders...
2024-04-04 22:29:17,756 - INFO - Processing worlds for erode map drawing
2024-04-04 22:29:22,159 - INFO - sorting routes...

4.4s - 2.6x